Introduce Cypher-based tutorial REPL
It basically imports a sequence of exercises at startup, in an
idempotent manner and saves progression in the configured graph
database.

The exercise expected result is a list of rows, i.e.
`List<Map<String,Object>>` serialized with Kryo and then
base 64-encoded.

The REPL also supports custom commands (prefixed with ':'),
scanned dynamically on the classpath. Default commands are:

 - `:commands` shows all commands
 - `:show` shows current exercise instructions
 - `:reset` resets user progression, back to first exercise
 - `:exit` quits REPL

There is also a standalone generator that helps in creating the
proper exercise Cypher descriptor file. The latter is to be
versioned within the workshop repository.

Messages are printed via `ConsoleLogger`, which directly relies
on JLine3 styling API for fun and profit.